refactored dtos structure
