+++ src/main/java/com/app/bookstore/dto/book/CreateBookRequestDto.java
@@ -1,4 +1,4 @@
-package com.app.bookstore.dto;
+package com.app.bookstore.dto.book;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
