+++ src/main/java/com/app/bookstore/dto/user/UserLoginResponseDto.java
@@ -1,4 +1,4 @@
-package com.app.bookstore.dto;
+package com.app.bookstore.dto.user;
 
 public record UserLoginResponseDto(String token) {
 }
