+++ src/main/java/com/app/bookstore/mapper/BookMapper.java
@@ -1,8 +1,8 @@
 package com.app.bookstore.mapper;
 
 import com.app.bookstore.config.MapperConfig;
-import com.app.bookstore.dto.BookDto;
-import com.app.bookstore.dto.CreateBookRequestDto;
+import com.app.bookstore.dto.book.BookDto;
+import com.app.bookstore.dto.book.CreateBookRequestDto;
 import com.app.bookstore.model.Book;
 import org.mapstruct.Mapper;
