+++ src/main/java/com/app/bookstore/service/book/BookService.java
@@ -1,7 +1,7 @@
 package com.app.bookstore.service.book;
 
-import com.app.bookstore.dto.BookDto;
-import com.app.bookstore.dto.CreateBookRequestDto;
+import com.app.bookstore.dto.book.BookDto;
+import com.app.bookstore.dto.book.CreateBookRequestDto;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
