+++ src/main/java/com/app/bookstore/service/user/UserService.java
@@ -1,7 +1,7 @@
 package com.app.bookstore.service.user;
 
-import com.app.bookstore.dto.UserRegistrationRequestDto;
-import com.app.bookstore.dto.UserResponseDto;
+import com.app.bookstore.dto.user.UserRegistrationRequestDto;
+import com.app.bookstore.dto.user.UserResponseDto;
 import com.app.bookstore.exception.RegistrationException;
 
 public interface UserService {
