+++ src/main/java/com/app/bookstore/service/user/UserServiceImpl.java
@@ -1,7 +1,7 @@
 package com.app.bookstore.service.user;
 
-import com.app.bookstore.dto.UserRegistrationRequestDto;
-import com.app.bookstore.dto.UserResponseDto;
+import com.app.bookstore.dto.user.UserRegistrationRequestDto;
+import com.app.bookstore.dto.user.UserResponseDto;
 import com.app.bookstore.exception.RegistrationException;
 import com.app.bookstore.mapper.UserMapper;
 import com.app.bookstore.model.User;
